+++ CHANGELOG.md
@@ -10,6 +10,14 @@
 
   The second (`scoped_metric`) and third (`elapsed`) arguments in the `NewRelic::Agent::Datastores.notice_sql` method are deprecated. They have not been used by the method for some time. Instead, these values will be set based on the current segment when the API is called. [PR#3345](https://github.com/newrelic/newrelic-ruby-agent/pull/3345)
 
+- **Feature: Deprecation notice for second argument in Datastores.notice_statement API**
+
+  The second (`elapsed`) argument in the `NewRelic::Agent::Datastores.notice_statement` method is deprecated. It has not been used by the method for some time. Instead, this value will be set based on the current segment when the API is called. [PR#3346](https://github.com/newrelic/newrelic-ruby-agent/pull/3346)
+
+- **Feature: Deprecation notice for proc's second and third arguments in Datastores.wrap API**
+
+  The `NewRelic::Agent::Datastores.wrap` method is changing. In a future major version, proc will only accept a single argument, the result of the yield. The scoped metric name and elapsed arguments will be removed, as they are being removed from the `Datastores.notice_sql` method. The scoped metric name and elapsed values are derived from the current segment when the wrap yields. [PR#3346](https://github.com/newrelic/newrelic-ruby-agent/pull/3346)
+
 - **Feature: Add Trace ID Ratio Based sampling options**
 
   The agent can now sample traces using the OpenTelemetry Trace ID Ratio Based sampler algorithm. This samples traces based on a probability between 0.0 and 1.0 based on the trace ID.